Fixed a composing containerizer hanging in `wait()` after recovery.
This patch fixes the issue when `wait()` of composing containerizer
returns a future that is never set to ready state due to a missing
`wait()` call on corresponding containerizer after it's recovery.
It is necessary to call `wait()` for each known container in composing
containerizer, because any terminated container should be handled in
`_destroy()` callback that is bound to `wait()`.