Removed `destroyed` from `Container` struct in composing containerizer.
Previously, we were using `destroyed` promise for each container to
guarantee that calling `destroy()` on a container returns a non-empty
value when the destroy-in-progress stops an launch-in-progress using
the next containerizer. Since we are unifying the semantics of the
return type for both `wait()` and `destroy()` operations, we can
remove the `destroyed` promise.

Review: https://reviews.apache.org/r/66668