Cleaned up container on launch failures in composing containerizer.
Previously, if a parent container was unknown to the composing
containerizer during an attempt to launch a nested container
via `ComposingContainerizerProcess::launch()`, the composing
containerizer returned an error without cleaning up the container.
The `containerizer` field was uninitialized, so a further attempt
to remove or destroy the nested container led to segfault.

This patch removes the container when the parent container is unknown.

Review: https://reviews.apache.org/r/68235/