Fixed the FD leak if containerizer::_launch() failed or discarded.
For the period between IOSB server is up and container process exec,
if the containerizer launch returns failure or discarded, the FD used
for signaling from the container process to the IOSB finish redirect
will be leaked, which would cause the IOSB stuck at `io::redirect`
forever. It would block the containerizer from cleaning up this
container at IOSB.

This issue could be exposed if there are frequent check containers
launch on an agent with heavy loads.

This patch fixes the issue by handling discard of a `launch` future,
so the container IO is cleaned up and therefore all FDs are closed.

Review: https://reviews.apache.org/r/69684