Modified `LOGROTATE_CustomRotateOptions` test after memfd change.
This patch changes the mechanism of verifying application of custom
logrotation options. Since we don't create logrotate config file
in the container's sandbox on Linux, we can't use the previous
approach where we tried to find the custom options in the generated
`stdout.logrotate.conf`. Instead, we inject a script into custom
options, which creates an empty file in the temporary directory on
log rotation.

Review: https://reviews.apache.org/r/70011